+++ examples/frameworks/__init__.py
@@ -1,0 +1,1 @@
+"""Framework-hosting recipes."""